# api/openapi.yaml
openapi: 3.0.4
info:
  title: Participium - OpenAPI 1.0
  description: Participium Server based.
  termsOfService: https://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.12
...
paths:
  /auth/users:
    post:
      tags:
        - Auth
      summary: Logs user into the system.
      description: Log into the system.
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: username
                password:
                  type: string
                  format: password
                  example: secure_word
              required:
                - username
                - password
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        "200":
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/xml:
              schema:
                type: string
            application/json:
              schema:
                type: string
        "400":
          description: Invalid username/password supplied
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /users:
    post:
      tags:
        - Users
      summary: Create user.
      description: This can only be done by the logged in user.
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
          application/xml:
            schema:
              $ref: "#/components/schemas/User"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
            application/xml:
              schema:
                $ref: "#/components/schemas/User"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/generateotp:
    post:
      tags:
        - Users
      summary: Generate OTP for user
      description: Generate a One-Time Password (OTP) for the user to enhance security during sensitive operations.
      operationId: generateOtp
      requestBody:
        description: User information for OTP generation
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@example.com
              required:
                - email
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  otp:
                    type: string
                    example: "123456"
        "403":
          description: Account is already active
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Account is already active"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"    
  /users/verifyotp:
    post:
      tags:
        - Users
      summary: Verify OTP for user
      description: Verify the One-Time Password (OTP) provided by the user to confirm their identity.
      operationId: verifyOtp
      requestBody:
        description: User information for OTP verification
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@example.com
                otp:
                  type: string
                  example: "123456"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean
                    example: true
        "401":
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid or expired OTP"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/me:
    get:
      tags:
        - Users
      summary: Retrieve authenticated user's profile
      description: Returns the profile of the authenticated user (avatar, telegramUsername, emailNotifications included).
      operationId: getMyProfile
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      tags:
        - Users
      summary: Update authenticated user's profile
      description: Update telegram username, email notification preference and optionally upload an avatar image.
      operationId: updateMyProfile
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
                telegramUsername:
                  type: string
                emailNotifications:
                  type: boolean
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/logout:
    get:
      tags:
        - Users
      summary: Logs out current logged in user session.
      description: Log user out of the system.
      operationId: logoutUser
      parameters: []
      responses:
        "200":
          description: successful operation
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        username:
          type: string
          example: username
        firstName:
          type: string
          example: Mario
        lastName:
          type: string
          example: Rossi
        password:
          type: string
          example: placeholder
        email:
          type: string
          example: mail@domain.com
        avatar:
          type: string
          example: /uploads/avatars/167123123-avatar.jpg
        telegramUsername:
          type: string
          example: my_telegram
        emailNotifications:
          type: boolean
          example: true
      xml:
        name: user
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header